ci(spread): only run docs tests if docs are changed
These tests take a disproportional amount of time to run, and don't need to run
on every code change.

This adds a separate job, 'docs-tests', that runs the spread tests for the docs
but only if a) a file in the docs/ dir is changed, or b) the pull request has
the 'run-manual-spread' label, or c) this is a scheduled run or a run in 'main'.